remove Client class state and move it to a new class: Client::Capabilities
