+++ lib/datadog/core/remote.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative 'remote/client'
-
 module Datadog
   module Core
     # Module to interact with core remote component
@@ -9,24 +7,6 @@ module Datadog
       class << self
         def active_remote
           remote
-        end
-
-        def register_capabilities(capabilities)
-          return unless active_remote
-
-          Client.register_capabilities(capabilities)
-        end
-
-        def register_receivers(receivers)
-          return unless active_remote
-
-          Client.register_receivers(receivers)
-        end
-
-        def register_products(products)
-          return unless active_remote
-
-          Client.register_products(products)
         end
 
         private
