+++ lib/datadog/core/remote/component.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'worker'
+require_relative 'client/capabilities'
 require_relative 'client'
 require_relative '../transport/http'
 require_relative '../remote'
@@ -19,9 +20,9 @@ module Datadog
 
           transport_v7 = Datadog::Core::Transport::HTTP.v7(**transport_options.dup)
 
-          Client.init
+          capabilities = Client::Capabilities.new(settings)
 
-          @client = Client.new(transport_v7)
+          @client = Client.new(transport_v7, capabilities)
           @worker = Worker.new(interval: settings.remote.poll_interval_seconds) { @client.sync }
         end
